+++ src/components/css/gradientText.ts
