+++ src/components/icons/exitIcon.ts
@@ -1,5 +1,5 @@
 import {svg} from 'lit';
-import {color} from '../utils/colors';
+import {color} from '../css/colors';
 
 export const exitIcon = svg`
 <svg width="17" height="16" viewBox="0 0 17 16" fill="none" xmlns="http://www.w3.org/2000/svg">
