+++ src/components/icons/lwcIcon.ts
@@ -1,5 +1,5 @@
 import {svg} from 'lit';
-import {color} from '../utils/colors';
+import {color} from '../css/colors';
 
 export const lwcIcon = svg`
 <svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
